Fixup build
