+++ apps/prairielearn/assets/scripts/esm-bundles/react-fragments/CreateQuestionModalContents.ts
@@ -1,4 +1,6 @@
+import '@prairielearn/preact/client-runtime';
+import { registerReactFragment } from '@prairielearn/preact/client';
+
 import { CreateQuestionModalContents } from '../../../../src/components/CreateQuestionModalContents.js';
-import { registerReactFragment } from '../../behaviors/react-fragments/index.js';
 
 registerReactFragment(CreateQuestionModalContents);
