+++ apps/prairielearn/assets/scripts/esm-bundles/react-fragments/InstructorAssessmentQuestionsTable.ts
@@ -1,4 +1,6 @@
+import '@prairielearn/preact/client-runtime';
+import { registerReactFragment } from '@prairielearn/preact/client';
+
 import { InstructorAssessmentQuestionsTable } from '../../../../src/pages/instructorAssessmentQuestions/components/InstructorAssessmentQuestionsTable.js';
-import { registerReactFragment } from '../../behaviors/react-fragments/index.js';
 
 registerReactFragment(InstructorAssessmentQuestionsTable);
