+++ apps/prairielearn/assets/scripts/esm-bundles/react-fragments/InstructorInstanceAdminBillingForm.ts
@@ -1,4 +1,6 @@
+import '@prairielearn/preact/client-runtime';
+import { registerReactFragment } from '@prairielearn/preact/client';
+
 import { InstructorInstanceAdminBillingForm } from '../../../../src/ee/pages/instructorInstanceAdminBilling/components/InstructorInstanceAdminBillingForm.js';
-import { registerReactFragment } from '../../behaviors/react-fragments/index.js';
 
 registerReactFragment(InstructorInstanceAdminBillingForm);
