+++ apps/prairielearn/assets/scripts/esm-bundles/react-fragments/InstructorStudentDetail.ts
@@ -1,4 +1,6 @@
+import '@prairielearn/preact/client-runtime';
+import { registerReactFragment } from '@prairielearn/preact/client';
+
 import { InstructorStudentDetail } from '../../../../src/pages/instructorStudentDetail/instructorStudentDetail.html.js';
-import { registerReactFragment } from '../../behaviors/react-fragments/index.js';
 
 registerReactFragment(InstructorStudentDetail);
