+++ apps/prairielearn/assets/scripts/esm-bundles/react-fragments/InstructorStudents.ts
@@ -1,4 +1,6 @@
+import '@prairielearn/preact/client-runtime';
+import { registerReactFragment } from '@prairielearn/preact/client';
+
 import { InstructorStudents } from '../../../../src/pages/instructorStudents/instructorStudents.html.js';
-import { registerReactFragment } from '../../behaviors/react-fragments/index.js';
 
 registerReactFragment(InstructorStudents);
