+++ packages/preact/src/client-runtime.tsx
@@ -1,24 +1,22 @@
-import './debug.js';
+/**
+ * This file, if imported, will register a selector observer that will hydrate
+ * React fragments on the client.
+ */
+if (process.env.NODE_ENV !== 'production') {
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
+  require('preact/debug');
+} else {
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
+  require('preact/devtools');
+}
 
-import { type ComponentType, hydrate } from 'preact';
 import { observe } from 'selector-observer';
 import superjson from 'superjson';
 
 import { onDocumentReady } from '@prairielearn/browser-utils';
+import { hydrate } from '@prairielearn/preact-cjs';
 
-import { ReactFragmentsRegistry } from './registry.js';
-
-const registry = new ReactFragmentsRegistry();
-
-export function registerReactFragment(component: ComponentType<any>, nameOverride?: string) {
-  // Each React component that will be hydrated on the page must be registered.
-  // Note that we don't try to use `component.name` since it can be minified or mangled.
-  const id = nameOverride ?? component.displayName;
-  if (!id) {
-    throw new Error('React fragment must have a displayName or nameOverride');
-  }
-  registry.setReactFragment(id, component);
-}
+import { registry } from './client.js';
 
 onDocumentReady(() => {
   observe('.js-react-fragment', {
